requirements.txt
playwright
transformers
chromadb
torch
python-dotenv
requests
fpdf